Don't use invalid arguments like --partial 29/28 at end of months.
Idea to use 28 chunks is to guarantee every domain gets refreshed at
least every 30 days.

If we use /31 then the domains on the last bucket would not be
refreshed on "short" months.

If we use /30 then the domains on the lasts buckets would not be
refreshed on february.

With /28 we're never scanning at the end of long months (let's call
this period "server holidays"?), we can a tad more on every days, but
most importantly we guarantee to scan each and every domain at least
once every calendar month.